Merge '20150712_new_types' into master.

+++ Source/Test/Src/DatabaseTestBase.cpp
@@ -122,12 +122,14 @@ BEGIN_NAMESPACE_DATABASE_TEST
 	STR( "	IDTest=?" );
 	String const QUERY_DIRECT_DELETE_ELEMENT = STR( "DELETE FROM Test WHERE IDTest=?" );
 
-	CDatabaseTest::CDatabaseTest( const String & type, const String & server, const String & database, const String & user, const String & password )
+	CDatabaseTest::CDatabaseTest( const String & type, const String & server, const String & database, const String & user, const String & password, bool hasNChar, bool hasSeparateBooleanAndTinyInt )
 		: _type( type )
 		, _server( server )
 		, _database( database )
 		, _user( user )
 		, _password( password )
+		, _hasNChar( hasNChar )
+		, _hasSeparateBooleanAndTinyInt( hasSeparateBooleanAndTinyInt )
 	{
 	}
 
@@ -137,10 +139,10 @@ BEGIN_NAMESPACE_DATABASE_TEST
 
 	boost::unit_test::test_suite * CDatabaseTest::Init_Test_Suite()
 	{
-		///@remarks Create the internal TS instance.
+		//!@remarks Create the internal TS instance.
 		testSuite = new boost::unit_test::test_suite( "CDatabaseSqliteTest" );
 
-		///@remarks Add the TC to the internal TS.
+		//!@remarks Add the TC to the internal TS.
 		testSuite->add( BOOST_TEST_CASE( std::bind( &CDatabaseTest::TestCase_CreateDatabase, this ) ) );
 		testSuite->add( BOOST_TEST_CASE( std::bind( &CDatabaseTest::TestCase_DatabaseQueryFieldsInsertRetrieve, this ) ) );
 		testSuite->add( BOOST_TEST_CASE( std::bind( &CDatabaseTest::TestCase_DatabaseQueryFieldsInsertRetrieveOtherIndex, this ) ) );
@@ -156,13 +158,13 @@ BEGIN_NAMESPACE_DATABASE_TEST
 		//testSuite->add( BOOST_TEST_CASE( std::bind( &CDatabaseTest::TestCase_DatabaseStatementStoredProcedure, this ) ) );
 		testSuite->add( BOOST_TEST_CASE( std::bind( &CDatabaseTest::TestCase_DestroyDatabase, this ) ) );
 
-		///@remarks Return the TS instance.
+		//!@remarks Return the TS instance.
 		return testSuite;
 	}
 
 	void CDatabaseTest::TestCase_CreateDatabase()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_CreateDatabase ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_CreateDatabase ****" );
 		{
 			auto const guard = make_block_guard( [this]()
 			{
@@ -177,104 +179,109 @@ BEGIN_NAMESPACE_DATABASE_TEST
 			{
 				DatabaseConnectionPtr connection = CreateConnection( *database, _server, _user, _password );
 
-				if ( connection && connection->IsConnected() )
+				if ( connection )
 				{
-					connection->CreateDatabase( _database );
-					connection->SelectDatabase( _database );
-					connection->ExecuteUpdate( _createTable );
+					if ( connection->IsConnected() )
+					{
+						connection->CreateDatabase( _database );
+						connection->SelectDatabase( _database );
+						connection->ExecuteUpdate( _createTable );
+					}
+
+					database->RemoveConnection();
 				}
 			}
 		}
-		CLogger::LogMessage( StringStream() << "**** End TestCase_CreateDatabase ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_CreateDatabase ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseQueryFieldsInsertRetrieve()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseQueryFieldsInsertRetrieve ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseQueryFieldsInsertRetrieve ****" );
 		TestCase_DatabaseFieldsInsertRetrieve< CDatabaseQuery >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseQueryFieldsInsertRetrieve ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseQueryFieldsInsertRetrieve ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseQueryFieldsInsertRetrieveOtherIndex()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseQueryFieldsInsertRetrieveOtherIndex ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseQueryFieldsInsertRetrieveOtherIndex ****" );
 		TestCase_DatabaseFieldsInsertRetrieveOtherIndex< CDatabaseQuery >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseQueryFieldsInsertRetrieveOtherIndex ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseQueryFieldsInsertRetrieveOtherIndex ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseQueryFieldsInsertRetrieveFast()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseQueryFieldsInsertRetrieveFast ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseQueryFieldsInsertRetrieveFast ****" );
 		TestCase_DatabaseFieldsInsertRetrieveFast< CDatabaseQuery >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseQueryFieldsInsertRetrieveFast ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseQueryFieldsInsertRetrieveFast ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseQueryFieldsInsertRetrieveFastOtherIndex()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseQueryFieldsInsertRetrieveFastOtherIndex ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseQueryFieldsInsertRetrieveFastOtherIndex ****" );
 		TestCase_DatabaseFieldsInsertRetrieveFastOtherIndex< CDatabaseQuery >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseQueryFieldsInsertRetrieveFastOtherIndex ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseQueryFieldsInsertRetrieveFastOtherIndex ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseQueryDirectQuery()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseQueryDirectQuery ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseQueryDirectQuery ****" );
 		TestCase_DatabaseDirectQuery< CDatabaseQuery >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseQueryDirectQuery ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseQueryDirectQuery ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseQueryStoredProcedure()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseQueryStoredProcedure ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseQueryStoredProcedure ****" );
 		TestCase_DatabaseStoredProcedure< CDatabaseQuery >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseQueryStoredProcedure ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseQueryStoredProcedure ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseStatementFieldsInsertRetrieve()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseStatementFieldsInsertRetrieve ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseStatementFieldsInsertRetrieve ****" );
 		TestCase_DatabaseFieldsInsertRetrieve< CDatabaseStatement >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseStatementFieldsInsertRetrieve ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseStatementFieldsInsertRetrieve ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseStatementFieldsInsertRetrieveOtherIndex()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseStatementFieldsInsertRetrieveOtherIndex ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseStatementFieldsInsertRetrieveOtherIndex ****" );
 		TestCase_DatabaseFieldsInsertRetrieveOtherIndex< CDatabaseStatement >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseStatementFieldsInsertRetrieveOtherIndex ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseStatementFieldsInsertRetrieveOtherIndex ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseStatementFieldsInsertRetrieveFast()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseStatementFieldsInsertRetrieve ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseStatementFieldsInsertRetrieve ****" );
 		TestCase_DatabaseFieldsInsertRetrieveFast< CDatabaseStatement >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseStatementFieldsInsertRetrieve ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseStatementFieldsInsertRetrieve ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseStatementFieldsInsertRetrieveFastOtherIndex()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseStatementFieldsInsertRetrieveFastOtherIndex ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseStatementFieldsInsertRetrieveFastOtherIndex ****" );
 		TestCase_DatabaseFieldsInsertRetrieveFastOtherIndex< CDatabaseStatement >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseStatementFieldsInsertRetrieveFastOtherIndex ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseStatementFieldsInsertRetrieveFastOtherIndex ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseStatementDirectQuery()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseStatementDirectQuery ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseStatementDirectQuery ****" );
 		TestCase_DatabaseDirectQuery< CDatabaseStatement >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseStatementDirectQuery ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseStatementDirectQuery ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseStatementStoredProcedure()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseStatementStoredProcedure ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseStatementStoredProcedure ****" );
 		TestCase_DatabaseStoredProcedure< CDatabaseStatement >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseStatementStoredProcedure ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseStatementStoredProcedure ****" );
 	}
 
 	void CDatabaseTest::TestCase_DestroyDatabase()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DestroyDatabase ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DestroyDatabase ****" );
 		{
 			auto const guard = make_block_guard( [this]()
 			{
@@ -289,35 +296,35 @@ BEGIN_NAMESPACE_DATABASE_TEST
 			{
 				DatabaseConnectionPtr connection = CreateConnection( *database, _server, _user, _password );
 
-				try
+				if ( connection )
 				{
-					connection->ExecuteUpdate( QUERY_DROP_TABLE );
-					connection->DestroyDatabase( _database );
-				}
-				catch ( std::exception & )
-				{
-					BOOST_CHECK( false );
-				}
+					if ( connection->IsConnected() )
+					{
+						connection->SelectDatabase( _database );
+						connection->ExecuteUpdate( QUERY_DROP_TABLE );
+						connection->DestroyDatabase( _database );
+					}
 
-				database->RemoveConnection();
+					database->RemoveConnection();
+				}
 			}
 		}
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DestroyDatabase ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DestroyDatabase ****" );
 	}
 
 #if defined( PERF_TEST )
 	void CDatabaseTest::TestCase_DatabaseStatementPerformances()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseStatementPerformances ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseStatementPerformances ****" );
 		TestCase_DatabasePerformances< CDatabaseQuery >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseStatementPerformances ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseStatementPerformances ****" );
 	}
 
 	void CDatabaseTest::TestCase_DatabaseQueryPerformances()
 	{
-		CLogger::LogMessage( StringStream() << "**** Start TestCase_DatabaseQueryPerformances ****" );
+		CLogger::LogInfo( StringStream() << "**** Start TestCase_DatabaseQueryPerformances ****" );
 		TestCase_DatabasePerformances< CDatabaseStatement >();
-		CLogger::LogMessage( StringStream() << "**** End TestCase_DatabaseQueryPerformances ****" );
+		CLogger::LogInfo( StringStream() << "**** End TestCase_DatabaseQueryPerformances ****" );
 	}
 #endif
